Disable syswow64 redirection just for `getFileVersionInfo` rather than for `_setProductInfo`
