+++ source/systemUtils.py
@@ -10,6 +10,17 @@ import winKernel
 import shellapi
 import winUser
 import os
+import functools
+import shlobj
+
+
+@functools.lru_cache(maxsize=1)
+def hasSyswow64Dir() -> bool:
+	"""Returns `True` if the current system has separate system32 directories for 32-bit processes."""
+	nativeSys32 = shlobj.SHGetKnownFolderPath(shlobj.FOLDERID.System.value)
+	Syswow64Sys32 = shlobj.SHGetKnownFolderPath(shlobj.FOLDERID.SystemX86.value)
+	return nativeSys32 != Syswow64Sys32
+
 
 def openUserConfigurationDirectory():
 	"""Opens directory containing config files for the current user"""
